# WinCCOAGraphQLRestApisExample/javascript/winccoa-graphql-server-main/restapi/openapi-full.yaml
openapi: 3.0.0
info:
  title: WinCC OA REST API
  version: 1.0.0
  description: |
    REST API for WinCC OA GraphQL Server - provides HTTP endpoints for all WinCC OA operations including:
    - Data point management (CRUD, value operations)
    - Data point type management
    - Tag queries with metadata
    - Alert handling
    - CNS (Central Navigation Service) operations
    - System information and redundancy
    - OPC UA configuration
  contact:
    name: API Support
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:4000
    description: Development server
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from /restapi/v1/auth/login
  schemas:
...
    Error:
      type: object
      properties:
        error:
          type: string
          example: Bad Request
        message:
          type: string
          example: Missing required parameter
...
    Tag:
      type: object
      properties:
        name:
          type: string
          example: ExampleDP_1.value
        value:
          description: Current tag value (any type)
          example: 42.5
        timestamp:
          type: string
          format: date-time
          example: "2024-01-01T12:00:00Z"
        status:
          type: object
          example: { "_online": { "_value": true } }
    TagValue:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
        value:
          description: Historical value (any type)
    TagHistory:
      type: object
      properties:
        name:
          type: string
        values:
          type: array
          items:
            $ref: '#/components/schemas/TagValue'
...
  responses:
    Unauthorized:
      description: Authentication required or token invalid
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: Unauthorized
            message: Missing or invalid authentication token
...
paths:
...
  /restapi/v1/tags:
    get:
      tags: [Tags]
      summary: Get tags
      description: Get multiple tags with value, timestamp, and status
      security:
        - bearerAuth: []
      parameters:
        - name: dpeNames
          in: query
          required: true
          schema:
            type: string
          description: Comma-separated list of data point element names
          example: "ExampleDP_1.value,ExampleDP_2.value"
      responses:
        '200':
          description: Tags retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  tags:
                    type: array
                    items:
                      $ref: '#/components/schemas/Tag'
  /restapi/v1/tags/history:
    get:
      tags: [Tags]
      summary: Get tag history
      description: Get historical data for multiple tags
      security:
        - bearerAuth: []
      parameters:
        - name: dpeNames
          in: query
          required: true
          schema:
            type: string
          description: Comma-separated list of data point element names
        - name: startTime
          in: query
          required: true
          schema:
            type: string
            format: date-time
        - name: endTime
          in: query
          required: true
          schema:
            type: string
            format: date-time
        - name: limit
          in: query
          schema:
            type: integer
        - name: offset
          in: query
          schema:
            type: integer
      responses:
        '200':
          description: Tag history retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  history:
                    type: array
                    items:
                      $ref: '#/components/schemas/TagHistory'

  # Alerts
...
  /restapi/v1/subscriptions/dp-connect:
    get:
      tags: [Subscriptions]
      summary: Subscribe to data point changes (SSE)
      description: |
        Opens a Server-Sent Events stream that emits a JSON event whenever the
        subscribed data point elements change.

        Wraps WinCC OA `dpConnect(callback, dpeNames, answer)`. The connection is
        cleaned up automatically when the client closes the stream.

        Each SSE event has the form:
        ```
        data: {"dpeNames":[...],"values":[...],"type":"...","error":null}
        ```

        To subscribe to multiple DPEs repeat the `dpeNames` parameter:
        `?dpeNames=Dp1:_online.._value&dpeNames=Dp2:_online.._value`
      security:
        - bearerAuth: []
      parameters:
        - name: dpeNames
          in: query
          required: true
          description: One or more data point element names to monitor. Repeat the parameter for multiple DPEs.
          style: form
          explode: true
          schema:
            oneOf:
              - type: string
              - type: array
                items:
                  type: string
          example: ExampleDP_Arg1:_online.._value
        - name: answer
          in: query
          required: false
          description: If true (default), send the current value immediately on connect in addition to future updates.
          schema:
            type: boolean
            default: true
      responses:
        '200':
          description: SSE stream of data point update events
          content:
            text/event-stream:
              schema:
                type: string
                description: |
                  Newline-delimited SSE stream. Each event is a `data:` line containing JSON:
                  { "dpeNames": ["dp:elem"], "values": [42], "type": "update", "error": null }
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /restapi/v1/subscriptions/dp-query-connect-single:
    get:
      tags: [Subscriptions]
      summary: Subscribe to query results — single (most recent) result set (SSE)
      description: |
        Opens a Server-Sent Events stream that emits a JSON event whenever the
        dpQuery result changes, returning only the most recent result set.

        Wraps WinCC OA `dpQueryConnectSingle(callback, answer, query, blockingTime)`.

        Each SSE event has the form:
        ```
        data: {"values":[[header,...],[row,...]],"type":"update","error":null}
        ```
      security:
        - bearerAuth: []
      parameters:
        - name: query
          in: query
          required: true
          description: dpQuery SQL-like query string.
          schema:
            type: string
          example: "SELECT '_original.._value' FROM 'ExampleDP_Arg*'"
        - name: answer
          in: query
          required: false
          description: If true (default), send the current result immediately on connect.
          schema:
            type: boolean
            default: true
        - name: blockingTime
          in: query
          required: false
          description: Blocking time in milliseconds for query updates.
          schema:
            type: number
      responses:
        '200':
          description: SSE stream of query result events
          content:
            text/event-stream:
              schema:
                type: string
                description: |
                  Newline-delimited SSE stream. Each event is a `data:` line containing JSON:
                  { "values": [["header"], ["row"]], "type": "update", "error": null }
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /restapi/v1/subscriptions/dp-query-connect-all:
    get:
      tags: [Subscriptions]
      summary: Subscribe to query results — complete result set on every update (SSE)
      description: |
        Opens a Server-Sent Events stream that emits a JSON event whenever the
        dpQuery result changes, returning the complete result set each time.

        Wraps WinCC OA `dpQueryConnectAll(callback, answer, query, blockingTime)`.

        Each SSE event has the form:
        ```
        data: {"values":[[header,...],[row,...]],"type":"update","error":null}
        ```
      security:
        - bearerAuth: []
      parameters:
        - name: query
          in: query
          required: true
          description: dpQuery SQL-like query string.
          schema:
            type: string
          example: "SELECT '_original.._value' FROM 'ExampleDP_Arg*'"
        - name: answer
          in: query
          required: false
          description: If true (default), send the current result immediately on connect.
          schema:
            type: boolean
            default: true
        - name: blockingTime
          in: query
          required: false
          description: Blocking time in milliseconds for query updates.
          schema:
            type: number
      responses:
        '200':
          description: SSE stream of query result events
          content:
            text/event-stream:
              schema:
                type: string
                description: |
                  Newline-delimited SSE stream. Each event is a `data:` line containing JSON:
                  { "values": [["header"], ["row"]], "type": "update", "error": null }
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'